Fix QML data structure version checking for ahead-of-time generated files
We must also do version checking for QML and JS files that were compiled
ahead of time and are embedded in resources. If the lookup for the
original source code fails, then we must generate an appropriate error
message.

As an upside we get better error reporting when trying to load an empty
file and Qt.include() now reports the error message in the statusText
field.

The error reporting for imported scripts was not changed as importing an
empty script is (oddly) allowed.

Task-number: QTBUG-66986
Change-Id: Ie0ef81af371a51ecf8c66ae7954d43f5cc6c12de
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>