Fix ABI/API symbol tagging
The findclasslist.pl perl script that produces the linker version script
got confused by the "struct name" that was part of a macro and thought
that the class "name" in the *_p.h was supposed to be annotated with the
private API tag, resulting in a "*4name*" mask in the linker script,
which in turn made lots of public symbols "private" that had name in it,
such as QQmlProperty::name(). Fixing the indentation works around it and
conforms to coding style.

Change-Id: I0c66a6bb1d49941d6ec6dd89d9433d9b6ae0c639
Done-with: Thiago Macieira <thiago.macieira@intel.com>
Task-number: QTBUG-67004
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>