Refactor QQuickStyleSelector
Instead of trying to determine a single base URL based on the import
URI of the plugin, and whether built in static mode, provide a list of
paths where to lookup styles, in priority order:

- requested style path (-style /path/to/style)
- QT_QUICK_CONTROLS_STYLE_PATH environment variable
- QT_INSTALL_QML/QtQuick/Controls.2/
- qrc://qt-project.org/imports/QtQuick/Controls.2/

Furthermore, provide the requested style name and the fallback style
name as a simple list of selectors. The lookup order is:

1) requested style
2) fallback style
3) default style

As a result, QQuickStyleSelector implementation is a lot simpler,
completely independent of QQuickStyle, and the style lookup works
regardless of whether the files are in QRC or on the file system.
This allows us to utilize QRC and the Qt Quick Compiler in the future.

Note: two data rows in tst_QQuickStyleSelector::select_data() had to
be fixed. Not sure how the expected values ended up like that, but as
the comments say, "Label.qml exists in the default and fallback styles"
and "Button.qml exists in all styles", so it makes no sense to expect
that Label.qml or Button.qml is selected from the FallbackStyle when
a valid "data" folder is specified.

Change-Id: I18dea9fddf8f079e0140b51b567814da0df2802c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>