Use AUX_QML_FILES for the Qt Quick Designer support files
The purpose of AUX_QML_FILES (handled in qml_module.prf) is to ensure
that certain files, usually provided for tooling purposes, are never
embedded into QRC, but always installed on the file system. Currently
it's only used for the plugins.qmltypes file, but it works perfectly
for the Qt Quick Designer support files too. None of these files should
ever be built into QRC, because it would only bloat the plugin, and
tools would not even see the files.

Change-Id: Id864a1753a7dd8c3729dd4b5113e6e7fe447ad4a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>