QQuickRangeSlider: add touchDragThreshold property
Add touchDragThreshold property for configuring the threshold to initiate a 'drag',
i.e. touch move, of the handle of the slider. The mouse 'drag' won't be affected
by the property.

This property may be also used to configure the drag events stealing priorities
of nested draggable items in the future.

[ChangeLog][RangeSlider] Added touchDragThreshold property for configuring the threshold
to initiate the touch 'drag' of the handle of the slider. The mouse 'drag' won't be
affected by the property.

Task-number: QTBUG-62784
Change-Id: I5555deb827de9d1dc7be00970fba15001105e419
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>