Add anchors property to Popup to allow centering in parent/window
Currently, users must manually position their Popup using x and y
bindings:

    Popup {
        x: (parent.width - width) / 2
        y: (parent.height - height) / 2
    }

This patch adds an anchors property so that you can do this instead:

    Popup {
        anchors.centerIn: parent
    }

It's also possible to conveniently center within the window from
anywhere within the scene (106e7b63 also documents an alternative way
of doing this using Overlay):

    Window {
        id: window

        Pane {
            Popup {
                anchors.centerIn: window
            }
        }
    }

QQuickAnchors were never used with Popup, because we cannot use the
QQuickAnchors implementation as-is, as the visual QQuickItem parent
is not the actual parent item of QQuickPopupItem.

Currently just centerIn is supported, as that's the most common
use case.

[ChangeLog][Controls][Popup] Added anchors.centerIn to Popup to allow
a covenient way of centering a popup.

Task-number: QTBUG-60354
Change-Id: Ia030f812df9da646fea8f373ef6199a21205ffbd
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>