QQuickPane: fix a regression in implicit size calculation
Caused by 87dec64. Previously, the bindings were:

    contentWidth: contentItem.implicitWidth || (contentChildren.length === 1 ? contentChildren[0].implicitWidth : 0)
    contentHeight: contentItem.implicitHeight || (contentChildren.length === 1 ? contentChildren[0].implicitHeight : 0)

That is, if contentItem does not have implicit size, Pane uses the
implicit size of the first content child (if there's one). When the
new automatic content size calculation was implemented, I forgot to
install a listener on the content child (in addition to contentItem).
Consequently, if the imlicit size of the first content child changed
after construction, the content size was not updated as appropriate.

Change-Id: Ib3d81f4e08a4d0822849d5db12429a2dd036e206
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>