Don't error on duplicate property names in object literals
This has been relaxed from ES5. In ES8 this is actually allowed
even in strict mode. According to the spec we are to evaluate all
rhs expressions, but assign the last one used in the object literal.

The spec probably required this relaxation to be able to handle
computed property names.

Change-Id: Ia1b02010b7541946029951b36e5457a07fdee818
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>