Fix Shortcuts in Popups with window context not being activated
Before this patch, this code wouldn't work:

    Popup {
        id: popup

        Shortcut {
            sequence: "Tab"
            onActivated: popup.visible = !popup.visible
        }
    }

Even though Popup is aware of its Window, the Shortcut object is
parented to Popup::contentItem, which does not have a Window associated
with it while the Popup is hidden.

The fix is to check if the item has a window in
QQuickShortcutContext::matcher(), so that that check fails and the
QQuickPopup itself is used to access the window.

Change-Id: Ia1c5d75ad6b82fb4c8b7664b6d418d84428b1ddf
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>