Fix QQuickControl::contentItemChange()
All other similar change() virtuals, such as fontChange(), localeChange()
and paletteChange(), are called _after_ the value has been applied. The
old value is passed as an argument. Use the same order for contentItem.

This allows calling such methods from contentItemChange(), that rely on
d->contentItem being set. If d->contentItem points to the old content
item, all methods need to take the new item as an argument, which gets
ugly especially if the same methods are used from elsewhere too. This
is very well visible in QQuickPanePrivate:

    - bool updateContentWidth(QQuickItem *item);
    - bool updateContentHeight(QQuickItem *item);
    - void updateContentSize(QQuickItem *item);

These are called with "newItem" from contentItemChange(), and with
"d->contentItem" from many other places.

Note: if a pre-notifications were required for something in the future,
it would be probably better to add "aboutToChange" counterparts.

Change-Id: I67dfdcce0e5f1cd9d9a37a63a0a9fdeaf82eafe6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>