Avoid marking hidden windows as updatePending in Gui render loop
Since eeb320bbd8763f3e72f79369cc3908e999a0da3c the GL context only
deletes textures when all windows with pending updates have finished
rendering.

renderWindow will not process any window that is not visible. This
leaves a logic bug that we can have the updatePending flag set but
never cleared.

If we have two windows, this leaves the other window still updating
normally, but lastDirtyWindow will always be false and we never call
endSync.

This results in an effective memory leak of all textures.

This patch resets the flag on hide() a move that can be considered safe
given the show() method will reset this flag anyway.

Change-Id: Iab0171716e27e31077a66b5e36a00bf28a2e7a8c
Reviewed-by: Kai Uwe Broulik <kde@privat.broulik.de>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>
Reviewed-by: Aleix Pol
Reviewed-by: Andy Nichols <andy.nichols@qt.io>