RangeSlider: add first.moved() and second.moved() signals
This is necessary to respond to changes in values, as the documentation
currently advises using onValuedChanged, which can result in
binding loop errors and is generally not the right way to respond
to input changes.

[ChangeLog][Controls][RangeSlider] Added a moved() signal to
each handle (similar to the Slider's moved() signal) to
react to the values being interactively changed by the user.

Task-number: QTBUG-67311
Change-Id: I4a026042855c69f22755415031bac8833cd566a9
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>