TextArea: fix background parent
QQuickTextAreaPrivate::attachFlickable() reparents the background from
Flickable::contentItem to the Flickable itself. This avoids that the
background scrolls together with the content. However, when deferred
execution was enabled, attachFlickable() ended up being called before
the background had been executed. If the TextArea has already been
attached to a Flickable when the background gets exeuted, set the
appropriate parent in setBackground().

Task-number: QTBUG-67334
Change-Id: I0847fca5861fdddeac7d47d47316b74bb31500cb
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>