Fix Text with ElideRight not being rendered when reparented
Consider the following example:

    Item {
        width: 100
        height: 30

        Text {
            width: parent ? parent.width : 0
            height: parent ? parent.height : 0
            elide: Text.ElideRight
            text: "wot"
        }
    }

When setting the Text item's parent to null, its explicit width and
height are set to 0. When restoring its parent (the Item), its
explicit width and height are set to 100 and 30 again, but the text
itself is still not rendered.

The cause can be seen here:

    if (!(widthChanged || widthMaximum) && !d->isLineLaidOutConnected()) {
        // only height has changed
        if (newGeometry.height() > oldGeometry.height()) {
            if (!d->heightExceeded) // Height is adequate and growing.
                goto geomChangeDone;

heightExceeded was false, because 30 > 12 (or whatever the implicit
height happened to be), so the text was not laid out again, even though
it went from having an explicit height of 0 to an explicit height of
30.

Fix the issue by only executing the goto if the old explicit height
wasn't 0.

Task-number: QTBUG-60328
Task-number: QTBUG-67145
Change-Id: I7f4d2f95bc95c850133ba91ac2d1a02c7ee159b6
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>