Fix plugins.qmltypes for QtQml installation in static builds
Mention plugins.qmltypes in AUX_QML_FILES to ensure that it is always
copied to the build directory. qml_module.prf usually takes care of this
by having QML_FILES in qmldir.files with INSTALLS += qmldir, but with
CONFIG += builtin_resources (set for static) that's not the case.
AUX_QML_FILES is the correct variable though.

Task-number: QTBUG-67600
Change-Id: I4291b942e14ca26749758c3511240f824288f07e
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>