Merge QQuickPageLayout to QQuickPage
The idea of a separate QQuickPageLayout falls a bit short, because as
a non-QObject it cannot e.g. emit signals when the headers and footers
change as a result of being destroyed. Next, we want to add implicit
header and footer sizes, which has the same problem. QQuickPageLayout
is currently used by QQuickPage and QQuickDialog. We are now moving the
functionality to QQuickPage, and will use QQuickPage in QQuickDialog
instead.

Change-Id: I869e6ba0056d424e617c48c2e4f591c75cde48e5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>