Use Los Angeles to represent PST8PDT, not Vancouver
This amends commit 2b8b7a162be52f8cd6c2bc39f498a1ddfb59dd68.

The very last test (15.9.5.43-0-8) in the ES test-suite relies on
constructing a time *just* outside the valid range, to provoke (and
check it gets) a RangeError.  However, it uses the current time-zone
offset in computing this; and the local time -> UTC conversion in the
Date constructor uses the zone information for the zone at the given
time.  The TZ db uses each zone's local mean solar time for the period
before a formal zone was introduced; and Vancouver is further West
than the nominal meridian for its zone.  The test code throws in an
hour's correction in case it's run in daylight-saving time, which
would put the test time, when run during standard time, an hour (and a
millisecond, rather than just the millisecond) outside the valid Date
range, so as to ensure it's still a ms outside when the test is run
during DST.  However, the 12m28s mean solar offset overcomes the 1ms
and puts the time back in range durign DST (but still 47m32.001s
outside during standard time, which is why this problem only surfaced
when North America entered summer time).  Fortunately, Los Angeles is
further East than the meridian in question, and uses the same zone.

Task-number: QTBUG-67010
Change-Id: I58fd7bf14366ebe5842b8abd7e6802c9c78ca27e
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>