Fix QQuickListViewPrivate::snapItemAt for lists with spacing
If lists had spacing, snapping would have gaps where snapItemAt would
return nullptr eventhough there were items above and below in the list.

I.e. if we have a list with items of size 100, and a spacing of 100, and
scroll to 110 (right below the first item), then the first item would fail
because this is false:

    itemTop+itemSize()/2 >= pos (50 >= 110)

And the next item (the one we should have snapped to), would fail because
this is false:

    itemTop-prevItemSize/2 < pos (200-100/2 < 110)

The meaning of first part was probably to see if the middle of the current
item is inside the view.

The meaning of the second part was probably to see if the middle of the
previous item is outside the view.

This patch refactors the code so it's more visible what's happening and
fixes the bug by taking the spacing into account when computing the position of
the halfway positions.

Task-number: QTBUG-59852
Change-Id: I60eec0d16e91d2f22d70b97d11bcde5d88ac2997
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>