QQmlDelegateModel: include QQDMIncubationTask errors with our errors
With the snippet in the referenced bug report (and the parent patch
applied), the output is:

    qrc:/main.qml:19:19: QML ListView: Error creating delegate

There should be more errors after this line, but since the delegate
itself didn't have any errors, we need to also check with the
incubation task. After doing so, the output becomes:

    qrc:/main.qml:19:19: QML ListView: Error creating delegate:
    qrc:/main.qml: Object destroyed during incubation

This adds important context (for developers and for users reporting
issues in the future) that was previously missing.

Task-number: QTBUG-49224
Change-Id: Ic7ac1a06c7dbdf3746f960d28908cc10f6ae86f5
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>