Replace Material.buttonHeight with Material.touchTarget
The visual height of a Material design Button is 36px in the normal
variant, and 32px in the dense variant. The Button has 6px vertical
insets to gain 48px touch target height. Rename the old buttonHeight
property as touchTarget to make room for the visual buttonHeight that
is needed when porting Button to use the newly introduced background
insets.

Change-Id: Ic2932f71ca75a626ff5e9b0ce9930e1b3f24cf00
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>