Only create CallContext objects for bindings that might be signal handlers
Creating the callcontext for all bindings causes a 15% performance
regression in both of the moving images benchmarks of qmlbench.

But in most cases, we know that a binding can't be a signal handler,
as those always have to start with 'on'. Take this into account
and avoid creating the context for most binding expressions.

Task-number: QTBUG-67782
Task-number: QTBUG-67783
Change-Id: I9a25cb916e374c7d03693e49646ca28853c6ba54
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>